# services/platform-api/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Runtime Platform API
  version: 1.0.0
paths:
...
  /api/v1/auth/me:
    get:
      responses:
        "200":
          description: current principal
          content:
            application/json:
              schema:
                type: object
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: login success
          content:
            application/json:
              schema:
                type: object
        "401":
          description: invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...